% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx_spr.R
\name{approx_spr}
\alias{approx_spr}
\title{\code{approx_spr} aproximates equilibrium SPR from mean f across patches}
\usage{
approx_spr(fish, fleet, f)
}
\arguments{
\item{f}{}
}
\value{
an aproximate spr
}
\description{
\code{approx_spr} aproximates equilibrium SPR from mean f across patches
}
