% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-density-gradient.R
\name{calc_density_gradient}
\alias{calc_density_gradient}
\title{calculate density gradient}
\usage{
calc_density_gradient(pop, y, num_patches, density_modifier,
  b0_buffer = 1.25)
}
\arguments{
\item{density_modifier}{}
}
\value{
a movement matrix modifier
}
\description{
Calculates a modifier for the distance based adult movement matrix based
on biomass
}
