% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_recruits.R
\name{calculate_recruits}
\alias{calculate_recruits}
\title{\code{calcualte_recruits} calculates recruits
as a function of spawning stock biomass under
various forms of density dependence}
\usage{
calculate_recruits(pop, fish, num_patches, patch_habitat = 1,
  phase = "burn", move_matrix, rec_devs = 0)
}
\arguments{
\item{phase}{}
}
\value{
recruits of age min_age in each patch
}
\description{
\code{calcualte_recruits} calculates recruits
as a function of spawning stock biomass under
various forms of density dependence
}
\examples{
\dontrun{
calculate_recruits(pop, fish, num_patches = 10, phase = 'grow')
}
}
