% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_target.R
\name{catch_target}
\alias{catch_target}
\title{\code{catch_target} finds the right effort to produce a target level of catch}
\usage{
catch_target(total_effort, target_catch, pop, num_patches, mpa, fleet,
  use = "opt", fish, prior_profits, year, burn_years)
}
\arguments{
\item{fleet}{}
}
\value{
a total effort to be distributed by the fleet model
}
\description{
\code{catch_target} finds the right effort to produce a target level of catch
}
