% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_effort.R
\name{determine_effort}
\alias{determine_effort}
\title{determine_effort}
\usage{
determine_effort(last_effort, fleet, fish, y, burn_years, pop, mpa,
  num_patches, effort_devs, profit_lags = 4, previous_max = NA)
}
\arguments{
\item{last_effort}{effort in the last time period}

\item{fleet}{the fleet object}

\item{fish}{the fish object}

\item{y}{the current year}

\item{burn_years}{the last year of the burn period}

\item{pop}{the population object}

\item{mpa}{the mpa object}

\item{num_patches}{the number of patches in the system}
}
\value{
the effort in the current year
}
\description{
\code{determine_effort} determines the effort in the current time period based
on different functional forms
}
