% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribute_fleet.R
\name{distribute_fleet}
\alias{distribute_fleet}
\title{\code{distribute_fleet} moves the fishing fleet around}
\usage{
distribute_fleet(pop, effort, prior_profits, year, burn_years, fleet,
  num_patches, mpa)
}
\arguments{
\item{fleet}{}
}
\value{
a redistributed vector of effort in space
}
\description{
\code{distribute_fleet} moves the fishing fleet around
}
\examples{
\dontrun{
  distribute_fleet(pop, effort, fleet)
  }
}
