% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-costs2.R
\name{estimate_costs2}
\alias{estimate_costs2}
\title{estimate costs that produce a given B/Bmsy open access}
\usage{
estimate_costs2(pars, fish, fleet, b_ref_oa, sim_years = 100,
  burn_years = 25, num_patches = 1, use = "fit", lags = 0,
  sprinkler = FALSE, mpa_habfactor = 1)
}
\arguments{
\item{lags}{}

\item{max_cr_ratio}{the maximum cost to revenue ratio}
}
\value{
an estimate of costs
}
\description{
estimate costs that produce a given B/Bmsy open access
}
