% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-msy.R
\name{estimate_msy}
\alias{estimate_msy}
\title{Title}
\usage{
estimate_msy(effort, fish, fleet, sim_years = 25, burn_years = 25,
  mpa_size = 0, mpa_year = 100, num_patches = 1, use = "fit",
  seed = 42)
}
\arguments{
\item{use}{}
}
\value{
an estiamte of MSY
}
\description{
Title
}
