% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lbspr.R
\name{fit_lbspr}
\alias{fit_lbspr}
\title{\code{fit_lbspr}}
\usage{
fit_lbspr(fish, fleet, length_comps)
}
\arguments{
\item{length_comps}{counts of samples by length bin}
}
\value{
lbspr fits
}
\description{
\code{fit_lbspr}
}
