% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-timeseries.R
\name{generate_timeseries}
\alias{generate_timeseries}
\title{Generate a distribution of values over time}
\usage{
generate_timeseries(thing, cv, ac, time, percent_slope = 0,
  min_thing = 0)
}
\arguments{
\item{thing}{the thing to be jittered}

\item{ac}{the degree of autocorrelation 0-1}

\item{time}{the length of time series to simulate}

\item{sigma}{the degree of variation}
}
\value{
a vector of length time
}
\description{
Generates a time series of values with lognormal errors and
some temporal autocorrelation
}
\examples{
\dontrun{
generate_timeseries(thing = 2, sigma = .1, ac = 0.5, time = 10)
}
}
