% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grow_and_die.R
\name{grow_and_die}
\alias{grow_and_die}
\title{\code{grow_and_die} runs growth mortality and fishing}
\usage{
grow_and_die(numbers, f, mpa, fish, fleet, y)
}
\arguments{
\item{y}{}
}
\value{
numbers and catch at age for all ages above recruits
}
\description{
\code{grow_and_die} runs growth mortality and fishing
}
\examples{
\dontrun{
grow_and_die(numbers, fish, fleet)
}
}
