% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_lbspr_lite.R
\name{lbspr_lite}
\alias{lbspr_lite}
\title{\code{lbspr_lite} runs a simplified version of LBSPR assuming selectivity is known}
\usage{
lbspr_lite(f, fish, fleet, length_comps, use = "find-f")
}
\arguments{
\item{f}{fishing mortality}

\item{fish}{a fish object containing life history data}

\item{fleet}{a fleet object containing fleet data}

\item{length_comps}{length composition data numbers by length bin}

\item{use}{set to find-f to use in optim}
}
\value{
either ss of length comps or SPR
}
\description{
\code{lbspr_lite} runs a simplified version of LBSPR assuming selectivity is known
}
