% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_to_age.R
\name{length_to_age}
\alias{length_to_age}
\title{length_to_age}
\usage{
length_to_age(length_samples, cv, k, linf, t0, max_age, min_age = 0,
  time_step = 1, aging_method = "vbk")
}
\arguments{
\item{max_age}{}
}
\value{
a data frame of estimated numbers at age
}
\description{
length_to_age
}
