% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_curve.R
\name{logistic_curve}
\alias{logistic_curve}
\title{calculate probabilities from a logistic curve}
\usage{
logistic_curve(values, fifty, delta)
}
\arguments{
\item{fifty}{50 percent selection}

\item{delta}{delta on top of 50 percent to calculate 95 percent}
}
\value{
probabilities at each value
}
\description{
calculate probabilities from a logistic curve
}
