% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_fish.R
\name{move_fish}
\alias{move_fish}
\title{\code{move_fish} moves fish via a movement function}
\usage{
move_fish(here_pop, num_patches, fish, move_matrix)
}
\arguments{
\item{here_pop}{}
}
\value{
relocated adult fish
}
\description{
\code{move_fish} moves fish via a movement function
}
