% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpa_counterfactual.R
\name{mpa_counterfactual}
\alias{mpa_counterfactual}
\title{run mpa experiment}
\usage{
mpa_counterfactual(fish, fleet, year_mpa, mpa_size, sim_years, num_patches,
  burn_years, sprinkler = FALSE, mpa_habfactor = 1, enviro = NA,
  enviro_strength = NA, rec_driver = "stochastic",
  random_mpas = FALSE, min_size = 1, simseed = 42)
}
\arguments{
\item{min_size}{}
}
\value{
a list with two experiments, with and without MPAs
}
\description{
run mpa experiment
}
