% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-spasm.R
\name{plot_spasm}
\alias{plot_spasm}
\title{Plot Spasm}
\usage{
plot_spasm(sim, type = "patch", font_size = 14)
}
\arguments{
\item{sim}{a simulation produced by sim_fishery}

\item{type}{the type of plot to produce, one of patch,totals, and doughnut}

\item{font_size}{base font size for plots}
}
\value{
a ggplot object
}
\description{
Plot Spasm
}
\examples{
\dontrun{

plot_spasm(sim, type = "patch")

}


}
