% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_catch_curve.R
\name{run_catch_curve}
\alias{run_catch_curve}
\title{\code{run_catch_curve} runs a weighted catch curve on spasm data}
\usage{
run_catch_curve(length_comps, fish)
}
\arguments{
\item{length_comps}{a data frame with columns length_bin and numbers}

\item{fish}{a spasm fish object containing life history}
}
\value{
total mortality estimated by catch curve, z
}
\description{
\code{run_catch_curve} runs a weighted catch curve on spasm data
}
