% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_lengths.R
\name{sample_lengths}
\alias{sample_lengths}
\title{sample_lenghts}
\usage{
sample_lengths(n_at_age, cv, k, linf, t0, time_step = 1,
  sample_type = "catch", percent_sampled = 0.1, sample_col = NA,
  linf_buffer = 10)
}
\arguments{
\item{percent_sampled}{}
}
\value{
a data frame of sampled length frequencies
}
\description{
sample_lenghts
}
