% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_fishery.R
\name{sim_fishery}
\alias{sim_fishery}
\title{\code{sim_fishery} simulates an age structured spatially explicit
model forward with fleets etc.}
\usage{
sim_fishery(fish, fleet, manager, num_patches = 10, sim_years = 1,
  burn_years = 25, crashed_pop = 0.001, random_mpas = F,
  enviro = NA, enviro_strength = 1, rec_driver = "stochastic",
  est_msy = F, time_step, max_window = 10, min_size = 1,
  mpa_habfactor = 1, sprinkler = FALSE, keep_burn = FALSE,
  tune_costs = FALSE)
}
\arguments{
\item{crashed_pop}{}
}
\value{
a pop object with population and catch trajectories
}
\description{
\code{sim_fishery} simulates an age structured spatially explicit
model forward with fleets etc.
}
\examples{
\dontrun{
sim_fishery(fish = fish, fleet = fleet,...)
}

}
