% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_sampling.R
\name{sim_sampling}
\alias{sim_sampling}
\title{sim_sampling}
\usage{
sim_sampling(fish, fleet, sim_years = 25, burn_year = 25,
  percent_sampled = 0.25, num_patches = 1)
}
\arguments{
\item{num_patches}{}
}
\value{
length and age samples
}
\description{
sim_sampling
}
