% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strain_dlas.R
\name{strain_dlas}
\alias{strain_dlas}
\title{\code{strain_dlas} test sensitivity of dlas to input parameters}
\usage{
strain_dlas(fish, fleet, linf_error = 1, vbk_error = 1, m_error = 1,
  l50_error = 1, sigma_r = 0, rec_ac = 0, cv_len = 0,
  year_sampled = 1)
}
\arguments{
\item{l50_error}{}
}
\value{
a data frame of error terms
}
\description{
\code{strain_dlas} test sensitivity of dlas to input parameters
}
