% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_for_depletion.R
\name{tune_for_depletion}
\alias{tune_for_depletion}
\title{Tune for Depletion}
\usage{
tune_for_depletion(control_variable, target_depletion, fish, fleet,
  sim_years, burn_years, num_patches)
}
\arguments{
\item{num_patches}{}
}
\value{
ss for depletion
}
\description{
tunes a control variable to achieve a target depletion level
}
