% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_fish.R
\name{update_fish}
\alias{update_fish}
\title{update_fish}
\usage{
update_fish(fish)
}
\arguments{
\item{fish}{}
}
\value{
an updated fish list
}
\description{
update_fish
}
\examples{
\dontrun{
update_fish(fish \%>\% mutate(linf = 222))
}
}
