% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_fleet.R
\name{update_fleet}
\alias{update_fleet}
\title{update_fleet}
\usage{
update_fleet(fleet, fish)
}
\arguments{
\item{fish}{}
}
\value{
an updated fish list
}
\description{
update_fleet
}
\examples{
\dontrun{
update_fleet(fleet)
}
}
